Fix PVS-Studio warning: implicit int->qreal conversions cause precision loss
V601: The bool type is implicitly cast to the integer type
V636: Expression was implicitly cast from 'int' type to 'double' type

CCBUG:393431
(cherry picked from commit ed51df9bc022a95250e3d2a084d88589b18e0d35)

+++ libs/flake/KoImageData_p.cpp
@@ -159,6 +159,6 @@ qint64 KoImageDataPrivate::generateKey(const QByteArray &bytes)
     qint64 answer = 1;
     const int max = qMin(8, bytes.count());
     for (int x = 0; x < max; ++x)
-        answer += bytes[x] << (8 * x);
+        answer += qint64(bytes[x] << (8 * x));
     return answer;
 }
